fix fps, add VOPManager

+++ src/TLibMain/TCore_Tick.h
@@ -11,8 +11,8 @@ namespace TsiU
 	class Tick
 	{
 	public:
-		virtual u32 GetTick()		= 0;
-		virtual u32 GetTickPerSec()	= 0;
+		virtual s64 GetTick()		= 0;
+		virtual s64 GetTickPerSec()	= 0;
 	};
 }
